+++ log-viewer/src/features/database/components/SOQLView.ts
@@ -120,25 +120,6 @@ export class SOQLView extends LitElement {
         table-layout: fixed;
         margin-bottom: 1rem;
       }
-
-      .dropdown-container {
-        box-sizing: border-box;
-        display: flex;
-        flex-flow: column nowrap;
-        align-items: flex-start;
-        justify-content: flex-start;
-
-        label {
-          display: block;
-          color: var(--vscode-descriptionForeground);
-          cursor: pointer;
-          font-size: calc(var(--vscode-font-size) * 0.9);
-          font-weight: 400;
-          line-height: 1.4;
-          margin-bottom: 4px;
-          user-select: none;
-        }
-      }
     `,
   ];
 
@@ -148,14 +129,17 @@ export class SOQLView extends LitElement {
       <database-section title="SOQL Statements" .dbLines="${this.soqlLines}"></database-section>
 
       <datagrid-filter-bar>
-        <div slot="filters" class="dropdown-container">
-          <label for="soql-groupby-dropdown">Group by</label>
-          <vs-select id="soql-groupby-dropdown" label="Group by" @change="${this._soqlGroupBy}">
-            <vscode-option>SOQL</vscode-option>
-            <vscode-option>Namespace</vscode-option>
-            <vscode-option>None</vscode-option>
-          </vs-select>
-        </div>
+        <vs-select
+          slot="filters"
+          id="soql-groupby-dropdown"
+          prefix="Group"
+          label="Group by"
+          @change="${this._soqlGroupBy}"
+        >
+          <vscode-option>SOQL</vscode-option>
+          <vscode-option>Namespace</vscode-option>
+          <vscode-option>None</vscode-option>
+        </vs-select>
 
         <div slot="actions">
           <vscode-toolbar-button
@@ -354,14 +338,12 @@ export class SOQLView extends LitElement {
         headerTooltip: true,
         headerWordWrap: true,
       },
-      headerSortElement: function (column, dir) {
+      headerSortElement: function (_column, dir) {
         switch (dir) {
           case 'asc':
             return "<div class='sort-by--top'></div>";
-            break;
           case 'desc':
             return "<div class='sort-by--bottom'></div>";
-            break;
           default:
             return "<div class='sort-by'><div class='sort-by--top'></div><div class='sort-by--bottom'></div></div>";
         }
@@ -411,7 +393,7 @@ export class SOQLView extends LitElement {
 
             return (aRowData.relativeCost || 0) - (bRowData.relativeCost || 0);
           },
-          tooltip: function (e, cell, _onRendered) {
+          tooltip: function (_e, cell, _onRendered) {
             const { isSelective, relativeCost } = cell.getData() as GridSOQLData;
             let title;
             if (isSelective === null) {
@@ -510,7 +492,7 @@ export class SOQLView extends LitElement {
       },
     });
 
-    this.soqlTable.on('groupClick', (e: UIEvent, group: GroupComponent) => {
+    this.soqlTable.on('groupClick', (_e: UIEvent, group: GroupComponent) => {
       const { type } = window.getSelection() ?? {};
       if (type === 'Range') {
         return;
@@ -528,7 +510,7 @@ export class SOQLView extends LitElement {
       }
     });
 
-    this.soqlTable.on('rowClick', function (e, row) {
+    this.soqlTable.on('rowClick', function (_e, row) {
       const { type } = window.getSelection() ?? {};
       if (type === 'Range') {
         return;